fixed some stuff

+++ src/main/java/com/finalproject/finalproject/service/PostService.java
@@ -10,9 +10,11 @@ import com.finalproject.finalproject.model.repositories.CategoryRepository;
 import com.finalproject.finalproject.model.repositories.CityRepository;
 import com.finalproject.finalproject.model.repositories.PostRepository;
 import com.finalproject.finalproject.model.repositories.UserRepository;
+import com.finalproject.finalproject.utility.UserUtility;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.time.LocalDateTime;
@@ -34,33 +36,44 @@ public class PostService {
     @Autowired
     CityRepository cityRepository;
 
+    @Transactional
     public PostResponseDTO createPost(int id, PostDTO postDTO) {
 
         User user = userRepository.findById(id).orElse(null);
+        if (user == null){
+            throw new BadRequestException("No user with that id");
+        }
+        if ( postDTO.getDescription() == null ||  postDTO.getDescription().isBlank()){
+            throw new BadRequestException("No description");
+        }
         Post post = new Post();
+        if (!UserUtility.userHasCategory(user,postDTO.getCategoryName())) {
+            throw new BadRequestException("This user isn't qualified for that kind of work");
+        }
+        post.setCategory(categoryRepository.findByCategoryName(postDTO.getCategoryName()));
         post.setOwner(user);
         post.setDescription(postDTO.getDescription());
-        if (categoryRepository.existsByCategoryName(postDTO.getCategoryName())) {
-            post.setCategory(categoryRepository.getByCategoryName(postDTO.getCategoryName()));
+        if (cityRepository.findByCityName(postDTO.getCityName()) == null){
+            throw new BadRequestException("No city with that name");
         }
-        City city = cityRepository.findByCityName(postDTO.getCityName());
         post.setPostedDate(LocalDateTime.now());
-        post.setCity(city);
+        post.setCity(cityRepository.findByCityName(postDTO.getCityName()));
         post = postRepository.save(post);
-        user.getPosts().add(post);
-        userRepository.save(user);
         return  modelMapper.map(post, PostResponseDTO.class);
     }
 
-    public PostDTO deletePost(int id){
+    public PostResponseDTO deletePost(int id){
+        if (!postRepository.findById(id).isPresent()){
+            throw new BadRequestException("No post with that id");
+        }
         Post post = postRepository.deleteById(id);
-        PostDTO dto = modelMapper.map(post,PostDTO.class);
+        PostResponseDTO dto = modelMapper.map(post,PostResponseDTO.class);
         return dto;
     }
 
-    public List<PostDTO> getAllPosts() {
+    public List<PostResponseDTO> getAllPosts() {
         List<Post> posts = postRepository.findAll();
-        List<PostDTO> dtos = posts.stream().map(post -> modelMapper.map(post,PostDTO.class)).collect(Collectors.toList());
+        List<PostResponseDTO> dtos = posts.stream().map(post -> modelMapper.map(post,PostResponseDTO.class)).collect(Collectors.toList());
         return dtos;
     }
 }
